refactor: 타입 구조 변경에 따른 수정사항 반영
